[ipc] refactoring the security/crypto stuff to its own common file.
