+++ cmd/agent/api/common/security.go
@@ -1,0 +1,94 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2017 Datadog, Inc.
+
+package common
+
+import (
+	"crypto/rand"
+	"crypto/rsa"
+	"crypto/x509"
+	"crypto/x509/pkix"
+	"encoding/pem"
+	"math/big"
+
+	"fmt"
+	"net"
+	"time"
+)
+
+// GenerateKeyPair create a public/private keypair
+func GenerateKeyPair(bits int) (*rsa.PrivateKey, error) {
+	privKey, err := rsa.GenerateKey(rand.Reader, bits)
+	if err != nil {
+		return nil, fmt.Errorf("generating random key: %v", err)
+	}
+
+	return privKey, nil
+}
+
+// CertTemplate create x509 certificate template
+func CertTemplate() (*x509.Certificate, error) {
+	serialNumberLimit := new(big.Int).Lsh(big.NewInt(1), 128)
+	serialNumber, err := rand.Int(rand.Reader, serialNumberLimit)
+	if err != nil {
+		return nil, fmt.Errorf("failed to generate serial number: %s", err)
+	}
+
+	notBefore := time.Now()
+	notAfter := notBefore.Add(10 * 365 * 24 * time.Hour)
+	template := x509.Certificate{
+		SerialNumber: serialNumber,
+		Subject: pkix.Name{
+			Organization: []string{"Datadoc, Inc."},
+		},
+		NotBefore:             notBefore,
+		NotAfter:              notAfter,
+		BasicConstraintsValid: true,
+	}
+
+	return &template, nil
+}
+
+// GenerateRootCert generates a root certificate
+func GenerateRootCert(hosts []string, bits int) (
+	cert *x509.Certificate, certPEM []byte, rootKey *rsa.PrivateKey, err error) {
+
+	rootCertTmpl, err := CertTemplate()
+	if err != nil {
+		return
+	}
+
+	rootKey, err = GenerateKeyPair(bits)
+	if err != nil {
+		return
+	}
+
+	// describe what the certificate will be used for
+	rootCertTmpl.IsCA = true
+	rootCertTmpl.KeyUsage = x509.KeyUsageCertSign | x509.KeyUsageDigitalSignature
+	rootCertTmpl.ExtKeyUsage = []x509.ExtKeyUsage{x509.ExtKeyUsageServerAuth, x509.ExtKeyUsageClientAuth}
+
+	for _, h := range hosts {
+		if ip := net.ParseIP(h); ip != nil {
+			rootCertTmpl.IPAddresses = append(rootCertTmpl.IPAddresses, ip)
+		} else {
+			rootCertTmpl.DNSNames = append(rootCertTmpl.DNSNames, h)
+		}
+	}
+
+	certDER, err := x509.CreateCertificate(rand.Reader, rootCertTmpl, rootCertTmpl, &rootKey.PublicKey, rootKey)
+	if err != nil {
+		return
+	}
+	// parse the resulting certificate so we can use it again
+	cert, err = x509.ParseCertificate(certDER)
+	if err != nil {
+		return
+	}
+	// PEM encode the certificate (this is a standard TLS encoding)
+	b := pem.Block{Type: "CERTIFICATE", Bytes: certDER}
+	certPEM = pem.EncodeToMemory(&b)
+	return
+}
