+++ cmd/agent/api/server.go
@@ -11,18 +11,13 @@ sending commands and receiving infos.
 package api
 
 import (
-	"crypto/rand"
-	"crypto/rsa"
 	"crypto/tls"
 	"crypto/x509"
-	"crypto/x509/pkix"
 	"encoding/pem"
 	"fmt"
 	stdLog "log"
-	"math/big"
 	"net"
 	"net/http"
-	"time"
 
 	"github.com/DataDog/datadog-agent/cmd/agent/api/agent"
 	"github.com/DataDog/datadog-agent/cmd/agent/api/check"
@@ -34,81 +29,6 @@ import (
 var (
 	listener net.Listener
 )
-
-// GenerateKeyPair create a public/private keypair
-func GenerateKeyPair(bits int) (*rsa.PrivateKey, error) {
-	privKey, err := rsa.GenerateKey(rand.Reader, bits)
-	if err != nil {
-		return nil, fmt.Errorf("generating random key: %v", err)
-	}
-
-	return privKey, nil
-}
-
-// CertTemplate create x509 certificate template
-func CertTemplate() (*x509.Certificate, error) {
-	serialNumberLimit := new(big.Int).Lsh(big.NewInt(1), 128)
-	serialNumber, err := rand.Int(rand.Reader, serialNumberLimit)
-	if err != nil {
-		return nil, fmt.Errorf("failed to generate serial number: %s", err)
-	}
-
-	notBefore := time.Now()
-	notAfter := notBefore.Add(10 * 365 * 24 * time.Hour)
-	template := x509.Certificate{
-		SerialNumber: serialNumber,
-		Subject: pkix.Name{
-			Organization: []string{"Datadoc, Inc."},
-		},
-		NotBefore:             notBefore,
-		NotAfter:              notAfter,
-		BasicConstraintsValid: true,
-	}
-
-	return &template, nil
-}
-
-// GenerateRootCert generates a root certificate
-func GenerateRootCert(hosts []string, bits int) (
-	cert *x509.Certificate, certPEM []byte, rootKey *rsa.PrivateKey, err error) {
-
-	rootCertTmpl, err := CertTemplate()
-	if err != nil {
-		return
-	}
-
-	rootKey, err = GenerateKeyPair(bits)
-	if err != nil {
-		return
-	}
-
-	// describe what the certificate will be used for
-	rootCertTmpl.IsCA = true
-	rootCertTmpl.KeyUsage = x509.KeyUsageCertSign | x509.KeyUsageDigitalSignature
-	rootCertTmpl.ExtKeyUsage = []x509.ExtKeyUsage{x509.ExtKeyUsageServerAuth, x509.ExtKeyUsageClientAuth}
-
-	for _, h := range hosts {
-		if ip := net.ParseIP(h); ip != nil {
-			rootCertTmpl.IPAddresses = append(rootCertTmpl.IPAddresses, ip)
-		} else {
-			rootCertTmpl.DNSNames = append(rootCertTmpl.DNSNames, h)
-		}
-	}
-
-	certDER, err := x509.CreateCertificate(rand.Reader, rootCertTmpl, rootCertTmpl, &rootKey.PublicKey, rootKey)
-	if err != nil {
-		return
-	}
-	// parse the resulting certificate so we can use it again
-	cert, err = x509.ParseCertificate(certDER)
-	if err != nil {
-		return
-	}
-	// PEM encode the certificate (this is a standard TLS encoding)
-	b := pem.Block{Type: "CERTIFICATE", Bytes: certDER}
-	certPEM = pem.EncodeToMemory(&b)
-	return
-}
 
 // StartServer creates the router and starts the HTTP server
 func StartServer() {
@@ -131,7 +51,7 @@ func StartServer() {
 
 	//
 	hosts := []string{"127.0.0.1", "localhost"}
-	_, rootCertPEM, rootKey, err := GenerateRootCert(hosts, 2048)
+	_, rootCertPEM, rootKey, err := common.GenerateRootCert(hosts, 2048)
 	if err != nil {
 		stdLog.Fatalf("unable to start TLS server")
 		return
